A recursive sequence needs .Defer.Publish.
The sequence referencing the mutable variable which breaks the recursive
loop needs to be created with Observable.Defer.Publish so we have
control over when the mutable is read.

Since I want tokenSource to be multicast anyway, choose to break at that
point. In this case I am using .Replay(1) rather than .Publish as I need
replay semantics.